# [PERF] Regex: stop calling Simplify