@bwplotka 2025-09-22
What about allocations?