@bboreham 2025-09-22
There’s (amortised) zero allocations during regex execution for the FastRegexMatcher benchmark. 
The NewMatcher one goes down a bit. 

(On my phone so I don’t have the numbers to hand)