@beorn7 2025-10-01
Maybe consider adding a comment that calling Simplify here was tried in the past and was proven to be counterproductive.