@bboreham 2025-10-02
We could have a test that checks allocations are below a certain level for NewMatcher?